refactor: create invoices-receipts section
